+++ test-magnolia/shared/src/main/scala-2.12-2.13/zio/test/magnolia/DeriveGen.scala
@@ -16,9 +16,9 @@
 
 package zio.test.magnolia
 
+import _root_.zio.test.{Gen, Sized}
 import magnolia._
 import zio.random.Random
-import _root_.zio.test.{Gen, Sized}
 
 import java.time.{Instant, LocalDate, LocalDateTime}
 import java.util.UUID
