+++ test-magnolia/shared/src/main/scala-2.12-2.13/zio/test/magnolia/diff/DeriveDiff.scala
@@ -1,7 +1,7 @@
 package zio.test.magnolia.diff
 
 import magnolia._
-import zio.test._
+import zio.test.diff.{Diff, DiffResult}
 
 trait DeriveDiff {
   type Typeclass[A] = Diff[A]
